Re-write NewEntity function to take advantage of `auto`/`decltype` instead of `std::enable_if`
The previous implementation was working but was slightly erroneous: It was assuming that
because an object was derived from the class `EntitySet` its constructor had a certain
number of parameters, and if a different object was derived from the class `EntityBlock`,
its constructor had a fixed and different number of parameters. This was working because
the few classes that did not match this pattern were not calling this function, but this
was likely going to create problems in the future if this code was modified and the
person modifying it was not aware of the limitation of the previous implementation.

Instead of relying on this partially wrong pattern, the new implementation checks that
the constructor of the calling entity has the correct signature using the `auto`/`decltype`
pattern.